Move dict ops to dict_ops.c; declare str ops
